+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+## v0.1.1 - Problem Solution and Constraints
+* **Commit**: 2026-02-01 15:59
+* Update Problem constructor
+* class: Constraint, Solution
+* types: ConstraintFn, Penalty, Score
+* types: ObjectiveFn, SolutionCoreFn, SolutionStringFn, SolutionDisplayFn
 ## v0.1.0 - Discrete Optimization Problem
 * **Commit**: 2026-02-01 14:57
-* Problem class
-* Variable, Value types
-* Goal, ProblemType enums
+* class: Problem
+* types: Variable, Value 
+* enums: Goal, ProblemType